less duplicate code in pull- & query-operators.ts WARNING: return value of query-operators changed
